#1294 :truck: move testdatafactory to correct directory
